+++ src/rpc/core_rpc_server.cpp
@@ -52,6 +52,8 @@ using namespace epee;
 #include "net/parse.h"
 #include "crypto/hash.h"
 #include "rpc/archival_claim_source.h"
+#include "rpc/archival_shard_coverage.h"
+#include "rpc/archival_shard_fetch.h"
 #include "rpc/rpc_args.h"
 #include "rpc/rpc_handler.h"
 #include "core_rpc_server_error_codes.h"
@@ -1645,6 +1647,51 @@ namespace cryptonote
       return false;
     }
 
+    res.status = CORE_RPC_STATUS_OK;
+    return true;
+  }
+  //------------------------------------------------------------------------------------------------------------------------------
+  bool core_rpc_server::on_get_archival_shard_coverage(const COMMAND_RPC_GET_ARCHIVAL_SHARD_COVERAGE::request& /*req*/, COMMAND_RPC_GET_ARCHIVAL_SHARD_COVERAGE::response& res, epee::json_rpc::error& error_resp, const connection_context *ctx)
+  {
+    RPC_TRACKER(get_archival_shard_coverage);
+    (void)ctx;
+    try
+    {
+      auto& db = m_core.get_blockchain_storage().get_db();
+      db_rtxn_guard rtxn_guard(&db);
+      rpc::fill_archival_shard_coverage(db, res);
+    }
+    catch (const std::exception& e)
+    {
+      MERROR("Failed to gather archival shard coverage: " << e.what());
+      error_resp.code = CORE_RPC_ERROR_CODE_INTERNAL_ERROR;
+      error_resp.message = "Failed to gather archival shard coverage";
+      return false;
+    }
+    res.status = CORE_RPC_STATUS_OK;
+    return true;
+  }
+  //------------------------------------------------------------------------------------------------------------------------------
+  bool core_rpc_server::on_request_archival_shard(const COMMAND_RPC_REQUEST_ARCHIVAL_SHARD::request& req, COMMAND_RPC_REQUEST_ARCHIVAL_SHARD::response& res, epee::json_rpc::error& error_resp, const connection_context *ctx)
+  {
+    RPC_TRACKER(request_archival_shard);
+    (void)ctx;
+    try
+    {
+      if (!rpc::fill_request_archival_shard(req.shard_id, res))
+      {
+        error_resp.code = CORE_RPC_ERROR_CODE_WRONG_PARAM;
+        error_resp.message = "could not retrieve this archive";
+        return false;
+      }
+    }
+    catch (const std::exception& e)
+    {
+      MERROR("Failed to request archival shard: " << e.what());
+      error_resp.code = CORE_RPC_ERROR_CODE_INTERNAL_ERROR;
+      error_resp.message = "could not retrieve this archive";
+      return false;
+    }
     res.status = CORE_RPC_STATUS_OK;
     return true;
   }